feat: Add check if note exist before delete

+++ src/app/use-cases/notes/delete-note.ts
@@ -2,13 +2,13 @@ import { HTTP400Error } from "../../../errors/BadRequest.ts";
 import { ServerError } from "../../../errors/server-error.ts";
 import { prisma } from "../../../services/prisma.ts";
 
-export interface DeleteNoteUsecaseData{
+export interface DeleteNoteUsecaseData {
   owner: string;
   noteId: string;
 }
 
 export class DeleteNoteUsecase {
-  public execute = async ({owner, noteId}: DeleteNoteUsecaseData) => {
+  public execute = async ({ owner, noteId }: DeleteNoteUsecaseData) => {
     const isOwner = await prisma.user.findFirst({
       where: {
         id: owner
@@ -18,7 +18,13 @@ export class DeleteNoteUsecase {
     if (!isOwner) {
       throw new HTTP400Error('Owner invalido')
     }
-    
+
+    const isNote = await prisma.notes.findFirst({
+      where: { id: noteId }
+    })
+
+    if (!isNote) return;
+
     try {
       await prisma.notes.delete({
         where: {
@@ -27,8 +33,8 @@ export class DeleteNoteUsecase {
         }
       })
     } catch (error) {
-     console.error(error);
-     throw new ServerError('Não foi possivel deletar a nota');
+      console.error(error);
+      throw new ServerError('Não foi possivel deletar a nota');
     }
   }
 }
